fix MessagesHomePanel styling, delete MessagesInboxPage
