+++ assets/strings.en.json
@@ -1914,6 +1914,8 @@
     "panel.settings.profile.content.login.label": "Sign In/Sign Out",
     "panel.settings.profile.header.profile.label": "Profile",
 
+    "panel.settings.messages.header.messages.label": "Messages",
+
     "panel.settings.notifications.content.notifications.all.label": "All Notifications",
     "panel.settings.notifications.content.notifications.unread.label": "Unread Notifications",
     "panel.settings.notifications.content.preferences.label": "My Notification Preferences",
